Basic network call setup

+++ app/src/main/java/com/example/stock/repository/StockEntity.kt
@@ -2,12 +2,11 @@ package com.example.stock.repository
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import java.math.BigDecimal
 
 @Entity(tableName = "stock")
 data class StockEntity(
     @PrimaryKey val symbol: String,
     val name: String,
-    val price: BigDecimal,
+    val price: Double,
     val exchange: String
 )
